Added Models for Wattball Players
Athlete is the main model as Wattball Players inherits all fields
through Single Table Inheritance.

EmergencyContact is a simple model for storing players emergency
contact details, as they do not need a log in so don't need the User
table.

Address is a seperate table as it can be used in tickets as well,
just made sense to seperate it out.